changed ship speed type from size_t to int

+++ shm/Ship.cpp
@@ -31,5 +31,5 @@
     size_t Ship::getId() const        { return id_; }
     size_t Ship::getMaxCrew() const   { return maxCrew_; }
     std::string Ship::getName() const { return name_; }
-    size_t Ship::getSpeed() const     { return speed_; }
+    int Ship::getSpeed() const        { return speed_; }
